@@ -7,14 +7,15 @@
     "enzyme": "^3.7.0",
     "enzyme-adapter-react-16": "^1.7.0",
     "enzyme-to-json": "^3.3.4",
-    "react": "^16.6.3",
+    "moment": "^2.24.0",
+    "react": "^16.8.0",
     "react-body-classname": "^1.2.0",
-    "react-dom": "^16.6.3",
+    "react-dom": "^16.8.0",
     "react-onclickoutside": "^6.7.1",
     "react-redux": "^5.1.1",
     "react-router-dom": "^4.3.1",
     "react-router-redux": "^4.0.8",
-    "react-scripts": "2.1.1",
+    "react-scripts": "2.1.3",
     "redux": "^4.0.1",
     "redux-thunk": "^2.3.0",
     "styled-components": "^4.1.2"
