@@ -1,13 +1,22 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import RelatedArtists from 'components/ArtistPage/RelatedArtists/'
+import { RelatedArtists } from 'components/ArtistPage/RelatedArtists/'
 
 describe('Component | RelatedArtists', () => {
+  const relatedArtists = {
+    artists: [
+      {
+        id: '123',
+        name: 'artist name',
+        images: [{ url: 'path/to/photo', height: 200 }]
+      }
+    ]
+  }
   let component 
   beforeEach( () => {
     component = shallow(
-      <RelatedArtists />
+      <RelatedArtists relatedArtists={ relatedArtists } />
     )
   })
   it('should render a component', () => {
