@@ -1,13 +1,13 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import Album from 'components/ArtistPage/Albums/Album/'
+import StatelessArtistPage from 'components/ArtistPage/base'
 
-describe('Component | Album', () => {
+describe('Component | StatelessArtistPage', () => {
   let component 
   beforeEach( () => {
     component = shallow(
-      <Album id="123" title="album title" cover="path/to/image.png" releaseDate="2016-04-07" numberTracks={ 10 } />
+      <StatelessArtistPage />
     )
   })
   it('should render a component', () => {
