@@ -6,18 +6,16 @@ import PopUp from 'components/generic/PopUp/container'
 import Title from 'components/ArtistPage/Title/'
 import Album from 'components/ArtistPage/Albums/Album/'
 import StyledWrapperAlbums from 'components/ArtistPage/Albums/styles'
+import getImage from 'lib/get-image'
 
-const PopUpAlbums = ({ isOpen, togglePopUp }) => (
+const PopUpAlbums = ({ isOpen, togglePopUp, albums }) => (
   <PopUp isOpen={ isOpen } togglePopUp={ togglePopUp }>
     <Fragment>
       <Title color={ theme.blueDark } insidePopUp={ true }>Albums</Title>
       <StyledWrapperAlbums inPopup>
-        <Album title="Hallelujah" numberTracks={ 12 } releaseDate="2016-06-17" />
-        <Album title="Hallelujah" numberTracks={ 12 } releaseDate="2016-06-17" />
-        <Album title="Hallelujah" numberTracks={ 12 } releaseDate="2016-06-17" />
-        <Album title="Hallelujah" numberTracks={ 12 } releaseDate="2016-06-17" />
-        <Album title="Hallelujah" numberTracks={ 12 } releaseDate="2016-06-17" />
-        <Album title="Hallelujah" numberTracks={ 12 } releaseDate="2016-06-17" />
+        { albums.map( (album, index) => (
+          <Album key={ index } id={ album.id } title={ album.name } cover={ getImage(album.images) } numberTracks={ album.total_tracks } releaseDate={ album.release_date } />
+        )) }
       </StyledWrapperAlbums>
     </Fragment>
   </PopUp>      
@@ -25,7 +23,15 @@ const PopUpAlbums = ({ isOpen, togglePopUp }) => (
 
 PopUpAlbums.propTypes = {
   isOpen: PropTypes.bool,
-  togglePopUp: PropTypes.func.isRequired
+  togglePopUp: PropTypes.func.isRequired,
+  albums: PropTypes.arrayOf(
+    PropTypes.shape({
+      id: PropTypes.string,
+      name: PropTypes.string,
+      total_tracks: PropTypes.number,
+      release_date: PropTypes.string
+    })
+  )
 }
 
 export default PopUpAlbums
