@@ -1,42 +1,37 @@
 import React, { Fragment } from 'react'
 import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
 
 import WithContainer from 'components/generic/Container/'
+import getImage from 'lib/get-image'
 
-const BannerWrapper = ({
-  name,
-  photo,
-  genres
-}) => (
+export const BannerWrapper = ({ info }) => (
   <Fragment>
-    <img alt={ name } src={ photo } height="120" width="120" />
+    <img alt={ info.name } src={ getImage(info.images) } height="120" width="120" />
     <div>
-      <h1>{ name }</h1>
+      <h1>{ info.name }</h1>
       <ul>
-        { genres.map( (genre, index) => {
-          return <li key={ index }>{ genre } { index + 1 < genres.length ? <span>-</span> : null }</li>
+        { info.genres.map( (genre, index) => {
+          return <li key={ index }>{ genre } { index + 1 < info.genres.length ? <span>-</span> : null }</li>
         }) }
       </ul>
     </div>
   </Fragment>
 )
 
-BannerWrapper.defaultProps = {
-  name: 'Alicia Keys',
-  photo: 'https://i.scdn.co/image/05e25d030a1caa1611b1e7248d89b557d9867258',
-  genres: [
-    'hip pop',
-    'neo soul',
-    'pop',
-    'r&b',
-    'urban contemporary'
-  ]
-}
-
 BannerWrapper.propTypes = {
-  name: PropTypes.string.isRequired,
-  photo: PropTypes.string,
-  genres: PropTypes.array.isRequired
+  info: PropTypes.shape({
+    name: PropTypes.string,
+    images: PropTypes.array,
+    genres: PropTypes.array
+  }).isRequired
 }
 
-export default WithContainer(BannerWrapper)
+const mapStateToProps = (state) => ({
+  info: state.ArtistReducer.info
+})
+
+export default connect(
+  mapStateToProps,
+  {}
+)(WithContainer(BannerWrapper))
