@@ -1,4 +1,6 @@
 import React, { Component } from 'react'
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
 
 import Track from 'components/generic/Track/'
 import Title from 'components/ArtistPage/Title'
@@ -6,26 +8,50 @@ import Button from 'components/generic/Button/'
 import PopUpTopTracks from './PopUp/'
 import StyledWrapperTracks from './styles'
 
-class Tracks extends Component {
+export class Tracks extends Component {
   state = { isOpen: false }
   
   togglePopUp = () => this.setState({ isOpen: !this.state.isOpen })
  
   render () {
+    const { topTracks } = this.props
+    const topTracksList = topTracks && topTracks.tracks ? topTracks.tracks : []
     return (
       <div>
         <Title>Top Tracks</Title>
         <StyledWrapperTracks>
-          <Track name="Empire State of Mind" previewUrl="https://p.scdn.co/mp3-preview/d7624ec5f93b6d92c1836a95c40ecce463584f6e?cid=774b29d4f13844c495f206cafdad9c86" releaseDate="2016-07-17" />
-          <Track name="No one" previewUrl="https://p.scdn.co/mp3-preview/d7624ec5f93b6d92c1836a95c40ecce463584f6e?cid=774b29d4f13844c495f206cafdad9c86" releaseDate="2016-07-17" />
-          <Track name="Girl on Fire" previewUrl="https://p.scdn.co/mp3-preview/d7624ec5f93b6d92c1836a95c40ecce463584f6e?cid=774b29d4f13844c495f206cafdad9c86" releaseDate="2016-07-17" />
-          <Track name="Fallin'" previewUrl="https://p.scdn.co/mp3-preview/d7624ec5f93b6d92c1836a95c40ecce463584f6e?cid=774b29d4f13844c495f206cafdad9c86" releaseDate="2016-07-17" />
+          { topTracksList.map((track, index) => {
+            if ( index > 3 ) return null
+            return <Track key={ index } name={ track.name } previewUrl={ track.preview_url } releaseDate={ track.album.release_date } />
+          }) }
         </StyledWrapperTracks>
         <Button onClick={ this.togglePopUp }>view more tracks</Button>
-        <PopUpTopTracks isOpen={ this.state.isOpen } togglePopUp={ this.togglePopUp } />
+        <PopUpTopTracks isOpen={ this.state.isOpen } togglePopUp={ this.togglePopUp } tracks={ topTracksList } />
       </div>
     )
   }
 }
 
-export default Tracks
+Tracks.propTypes = {
+  topTracks: PropTypes.shape({
+    tracks: PropTypes.arrayOf(
+      PropTypes.shape({
+        name: PropTypes.string,
+        preview_url: PropTypes.string,
+        is_playable: PropTypes.bool,
+        album: PropTypes.shape({
+          release_date: PropTypes.string
+        })     
+      })
+    )
+  })
+}
+
+const mapStateToProps = (state) => ({
+  topTracks: state.ArtistReducer.topTracks
+})
+
+export default connect(
+  mapStateToProps,
+  {}
+)(Tracks)
